+++ Internal/TraceAndOptimizeBase/TraceAndOptimizeProcessor.cs
@@ -44,6 +44,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
         public bool SkipRemoveUnusedSubMesh;
         public bool SkipMergePhysBones;
         public bool SkipCompleteGraphToEntryExit;
+        public bool SkipReplaceEndBoneWithEndpointPosition;
         public bool SkipOptimizationWarnings;
 
         public Dictionary<SkinnedMeshRenderer, HashSet<string>> PreserveBlendShapes =
@@ -88,6 +89,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
             SkipRemoveUnusedSubMesh = config.debugOptions.skipRemoveUnusedSubMesh;
             SkipMergePhysBones = config.debugOptions.skipMergePhysBones;
             SkipCompleteGraphToEntryExit = config.debugOptions.skipCompleteGraphToEntryExit;
+            SkipReplaceEndBoneWithEndpointPosition = config.debugOptions.skipReplaceEndBoneWithEndpointPosition;
             SkipOptimizationWarnings = config.debugOptions.skipOptimizationWarnings;
 
             Enabled = true;