+++ CHANGELOG.md
@@ -38,6 +38,9 @@ The format is based on [Keep a Changelog].
   - This component will support both Skinned Mesh Renderer and Mesh Renderer.
 - Support for VRCSDK 3.9.1-beta `#1562`
   - New internal component ParentChangeDetector is added
+- Replace EndBone With Endpoint Position component which replaces the end bone in the vrc physbone with the Endpoint Position [`#1423`](https://github.com/anatawa12/AvatarOptimizer/pull/1423)
+- Automatic Replace EndBone With Endpoint Position [`#1423`](https://github.com/anatawa12/AvatarOptimizer/pull/1423)
+  - Trace and Optimize now automatically replaces PhysBone's EndBone with Endpoint Position if possible.
 
 ### Changed
 - Avatar Optimizer will run as late as possible in NDMF Pipeline by default `#1493`